refactor(auto_source): extract ContainerAssessor from LinkHeuristics
Own DOM observation (markers, headings, visible text, class/id tokens)
behind a deep container+anchor+destination_facts → ContainerSignals
seam so scoring stays on ContainerSignals and LinkHeuristics keeps
eligibility policy without co-locating observation helpers.
